hooks/eatmydata: download instead of copy-ing from host

+++ hooks/eatmydata/customize.sh
@@ -4,7 +4,22 @@ set -exu
 
 rootdir="$1"
 
-rm "$rootdir/usr/bin/eatmydata"
+if [ -e "$rootdir/var/lib/dpkg/arch" ]; then
+	chrootarch=$(head -1 "$rootdir/var/lib/dpkg/arch")
+else
+	chrootarch=$(dpkg --print-architecture)
+fi
+libdir="/usr/lib/$(dpkg-architecture -a $chrootarch -q DEB_HOST_MULTIARCH)"
+
+# if eatmydata was actually installed properly, then we are not removing
+# anything here
+if ! chroot "$rootdir" dpkg-query --list eatmydata; then
+	rm "$rootdir/usr/bin/eatmydata"
+fi
+if ! chroot "$rootdir" dpkg-query --list libeatmydata1; then
+	rm "$rootdir$libdir"/libeatmydata.so*
+fi
+
 mv "$rootdir/usr/bin/dpkg.orig" "$rootdir/usr/bin/dpkg"
 
 sync
